requirements.txt
Django==1.8.2
Pillow==2.9.0
wsgiref==0.1.2
